working dedicated window for hex viewer
