+++ Gui/Views/HexViewerWindow.axaml
@@ -1,0 +1,55 @@
+<Window
+	xmlns="https://github.com/avaloniaui"
+	xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+	xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
+	xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
+	xmlns:gui="using:OpenLoco.Gui"
+	xmlns:vm="using:OpenLoco.Gui.ViewModels"
+	mc:Ignorable="d" d:DesignWidth="800" d:DesignHeight="450"
+	x:Class="OpenLoco.Gui.Views.HexViewerWindow"
+	x:DataType="vm:HexWindowViewModel"
+	Title="HexViewerWindow"
+	WindowStartupLocation="CenterOwner"
+	SystemDecorations="Full"
+	Width="800"
+	Height="600"
+	ShowInTaskbar="False">
+
+	<DockPanel Background="{DynamicResource ExpanderContentBackground}">
+		<ScrollViewer>
+			<TreeView ItemsSource="{Binding CurrentHexAnnotations}" SelectedItem="{Binding CurrentlySelectedHexAnnotation}" Margin="4" DockPanel.Dock="Left" >
+				<TreeView.ItemTemplate>
+					<TreeDataTemplate ItemsSource="{Binding Nodes}">
+						<Grid>
+							<Grid.ColumnDefinitions>
+								<ColumnDefinition Width="Auto"/>
+								<ColumnDefinition Width="*"/>
+							</Grid.ColumnDefinitions>
+							<TextBlock Grid.Column="0" VerticalAlignment="Center" Text="{Binding Title}" Margin="2"/>
+							<TextBlock Grid.Column="1" VerticalAlignment="Center" HorizontalAlignment="Right" Text="{Binding OffsetText}" Margin="2" FontFamily="Courier New" />
+						</Grid>
+					</TreeDataTemplate>
+				</TreeView.ItemTemplate>
+			</TreeView>
+		</ScrollViewer>
+		<ScrollViewer>
+			<ItemsRepeater ItemsSource="{Binding CurrentHexDumpLines}" Background="{DynamicResource ButtonBackground}">
+				<ItemsRepeater.ItemTemplate>
+					<DataTemplate>
+						<Grid DockPanel.Dock="Left" Margin="4" x:DataType="vm:HexAnnotationLine">
+							<Grid.ColumnDefinitions>
+								<ColumnDefinition Width="Auto"/>
+								<ColumnDefinition Width="*"/>
+							</Grid.ColumnDefinitions>
+							<Border Grid.Column="0" Grid.ColumnSpan="1" Background="{DynamicResource ButtonBackgroundPointerOver}">
+								<TextBlock Text="{Binding Address}" VerticalAlignment="Center" Margin="4" FontFamily="Courier New" FontWeight="Bold" />
+
+							</Border>
+							<SelectableTextBlock Grid.Column="1" Margin="4" Text="{Binding Data}" VerticalAlignment="Center" FontFamily="Courier New" SelectionStart="{Binding SelectionStart}" SelectionEnd="{Binding SelectionEnd}" />
+						</Grid>
+					</DataTemplate>
+				</ItemsRepeater.ItemTemplate>
+			</ItemsRepeater>
+		</ScrollViewer>
+	</DockPanel>
+</Window>
